backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic>=2.10,<3.0
pydantic-settings>=2.7,<3.0
python-multipart>=0.0.20,<1.0
pypdf>=5.2,<6.0
python-docx>=1.1,<2.0
